day6: thread CharSet combiner through parsers
We need union or intersection, depending on which part we're solving (or
maybe others, who knows?).

But be careful: defaulting to List.foldl comb CharSet.empty where comb
is CharSet.intersection will just give the empty set! So we do my
standard trick of using the head of the list as the seed when available
(probably that should make its way over to lib/list).